refactor: ResourceReader 위치 변경
